removed use of ex to ref map, does gender matching on ref audio
Former-commit-id: 968b365bfd736d8b116e9ae43cdb87982b5aff5e

+++ src/main/java/mitll/langtest/server/database/analysis/IAnalysis.java
@@ -45,7 +45,7 @@ public interface IAnalysis {
 
   List<WordScore> getWordScoresForUser(long id, int minRecordings);
 
-  PhoneReport getPhonesForUser(long id, int minRecordings);
+  PhoneReport getPhonesForUser(int id, int minRecordings);
 
   List<UserInfo> getUserInfo(IUserDAO userDAO, int minRecordings);
 }
